adding new logic to Settings in RouterTunings
